feat(be-us-005): refactor user me endpoint
