+++ apps/backend/app/api/v1/auth.py
@@ -104,8 +104,3 @@ async def refresh(
     )
 
 
-@router.get("/me", response_model=UserResponse)
-async def get_me(
-    current_user: User = Depends(get_current_user),
-) -> UserResponse:
-    return UserResponse.model_validate(current_user)
