+++ apps/backend/app/api/v1/user.py
@@ -5,7 +5,7 @@ from fastapi import APIRouter, Depends, Query, status
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.core.database import get_db
-from app.core.dependencies import require_role
+from app.core.dependencies import get_current_user, require_role
 from app.enums.user_role import UserRole
 from app.models.user import User
 from app.schemas.user import UserResponse, UserUpdate
@@ -13,6 +13,13 @@ from app.services.user_service import UserService
 
 router = APIRouter()
 user_service = UserService()
+
+
+@router.get("/me", response_model=UserResponse)
+async def get_me(
+    current_user: User = Depends(get_current_user),
+) -> UserResponse:
+    return UserResponse.model_validate(current_user)
 
 
 @router.get("/", response_model=list[UserResponse], status_code=200)
